[treadmill] Make request type modular that we can easily extend
Summary:
Make request type modular so that we can easily add another workload by
implementing its request types in a separate source file. And I also
moved the request types of Memcached to MemcachedRequest.h and
MemcachedRequest.cpp as an example of how to do it.

Test Plan:
Compile it and run it against Memcached, and see no performance degradation
comparing to the previous implementation.

Reviewers: meisner

Reviewed By: meisner

Differential Revision: http://philz.eecs.umich.edu/D16