Get rid of most templating on Operation (only template by Request)
Summary:
Motivation: Thrift types in mcrouter implicitly encode the operation
associated with the request (e.g., McGetRequest represents a memcached get
operation).  To ease the coexistence of classic `McRequest`s and
`TypedThriftMessage<>` messages, we'll just encode the operation in the
class `McRequest` type by moving instances of `McRequest` to the new type
`McRequestWithOp`.

Reviewed By: pavlo-fb

Differential Revision: D2818206

fb-gh-sync-id: f670e71760587c18c4c8b63992dd7bb39014d6d0