Fix treadmill to properly use fb303 allowing response monitoring
Summary: We'd like to monitor treadmill response stats (e.g., How many 200s, 300s, etc). Right now the way we instantiate fb303 is broken and it doesn't export the stats the way we want. This fixes it and we get monitoring for free in the places where we bump counters.

Reviewed By: mchowX

Differential Revision: D4610232

fbshipit-source-id: 4cb830d26a36ec3fa050ee9093657a969a7e4b48