Remove to_send queue in Treadmill worker
Summary:
The Treadmill scheduler enqueues work to the to_send queue in the Treadmill worker with a fixed period, which the Treadmill worker drains as quickly as the server being tested can process requests.  If the Treadmill worker cannot keep up, the to_send queue grows without bound.

This has the effect of making it impossible to pause Treadmill. If there is a large amount of work in the to_send queue, and the scheduler is paused, the worker will continue sending requests as fast as it can until the to_send queue is drained.

Reviewed By: mchowX

Differential Revision: D5671828

fbshipit-source-id: ad7eab4bd7c66f1b2aadce84f413fc0127dd8961