add configuration methods
Summary:
Adds a configuration method to the treadmill thrift interface.  This is kept to a minimal map<string, string> to remain generic for the different treadmill clients.   As an initial implementation, the iris treadmill client adds a payload size.

This configuration is expected to be used by daemonized treadmill implementations to specify different parameters between runs.

Reviewed By: mchowX

Differential Revision: D15408477

fbshipit-source-id: 20084b50b7a778652f66cda5c1d6b37be82213eb