Refactor Treadmill.h to be extendable
Summary:
I need to change the worker constructor, but worker is constructed in this run function in the header file. Change it to be a class so that I can just override that part that construct the workers.
For more details about the changes I am going to make after this change : D24773671

Reviewed By: jdq

Differential Revision: D25100692

fbshipit-source-id: f5596cb58f979704df1352dbc386576ad86b0348